+++ app/components/core/MarkdownScreen.js
@@ -1,6 +1,5 @@
 import React, { Component } from "react";
 import PropTypes from 'prop-types';
-import * as R from 'ramda';
 import { View, Linking, Dimensions, Image, Text } from 'react-native';
 import { markdownStyle } from '../../themes/activityTheme';
 import { VideoPlayer } from './VideoPlayer';
@@ -9,17 +8,28 @@ import Markdown, { MarkdownIt, renderRules, tokensToAST, stringToTokens } from '
 import Mimoza from 'mimoza';
 import markdownContainer from 'markdown-it-container';
 import markdownIns from 'markdown-it-ins';
+import { WebView } from 'react-native-webview';
 
 const { width } = Dimensions.get('window');
 
-const markdownItInstance = MarkdownIt({typographer: true})
-.use(markdownContainer)
-.use(markdownContainer, 'hljs-left') /* align left */
-.use(markdownContainer, 'hljs-center')/* align center */
-.use(markdownContainer, 'hljs-right')/* align right */
-.use(markdownIns);
+const markdownItInstance = MarkdownIt({ typographer: true })
+  .use(markdownContainer)
+  .use(markdownContainer, 'hljs-left') /* align left */
+  .use(markdownContainer, 'hljs-center')/* align center */
+  .use(markdownContainer, 'hljs-right')/* align right */
+  .use(markdownIns);
+
+const regex = new RegExp(/^==(.*==)?/);
 
 const rules = {
+  text: (node, children, parent, styles, inheritedStyles = {}) => {
+    const additionalStyling = regex.test(node.content.trim()) ? { backgroundColor: 'yellow' } : {}
+    return (
+      <Text key={node.key} style={[inheritedStyles, styles.text, additionalStyling]}>
+        {checkNodeContent(node.content)}
+      </Text>
+    )
+  },
   image: (node, children, parent, styles, allowedImageHandlers, defaultImageHandler) => {
     const mimeType = Mimoza.getMimeType(node.attributes.src) || "";
 
@@ -33,7 +43,7 @@ const rules = {
           height={50}
         />
       );
-    } else if (mimeType.startsWith('video/')) {
+    } else if (mimeType.startsWith('video/') || node.attributes.src.includes('.quicktime')) {
       return (
         <View
           width={width - 20}
@@ -47,6 +57,22 @@ const rules = {
           />
         </View>
       );
+    } else if (node.attributes.src.includes('youtu')) {
+      let src = node.attributes.src.split(".be/")[1];
+      return (
+        <View
+          width={width - 20}
+          height={250}
+        >
+          <WebView
+            height={250}
+            key={node.key}
+            width={width - 20}
+            mediaPlaybackRequiresUserAction
+            source={{ uri: node.attributes.src.includes('watch?') ? node.attributes.src : `https://www.youtube.com/embed/${src}` }}
+          />
+        </View>
+      );
     }
 
     return (<Image
@@ -54,7 +80,7 @@ const rules = {
       style={{
         resizeMode: "contain",
         height: 200,
-        width: width-100
+        width: width - 100
       }}
       source={{
         uri: node.attributes.src
@@ -62,13 +88,19 @@ const rules = {
     />);
   },
   'container_hljs-left': (node, children, parent, styles) => {
-    return <View key={node.key}>{children}</View>
+    const style = 'flex-start';
+
+    return <View key={node.key} style={{ justifyContent: style, alignItems: style }}>{children}</View>
   },
   'container_hljs-center': (node, children, parent, styles) => {
-    return <View key={node.key}>{children}</View>
+    const style = 'center';
+
+    return <View key={node.key} style={{ justifyContent: style, alignItems: style }}>{children}</View>
   },
   'container_hljs-right': (node, children, parent, styles) => {
-    return (<View key={node.key}>{children}</View>);
+    const style = 'flex-end';
+
+    return (<View key={node.key} style={{ justifyContent: style, alignItems: style }}>{children}</View>);
   },
   'ins': (node, children) => {
     return (<Text key={node.key} style={{ textDecorationLine: 'underline' }}>{children}</Text>)
@@ -95,22 +127,32 @@ const rules = {
 }
 
 class MarkdownScreen extends Component {
-  shouldComponentUpdate(nextProps, nextState){
+  shouldComponentUpdate(nextProps, nextState) {
     return nextProps.children != this.props.children;
   }
 
   render() {
-    const { mstyle, children } = this.props;
+    let { mstyle, children } = this.props;
 
+    if (children.indexOf("404:") > -1) {
+      children = '# ¯\\\\_(ツ)_/¯ ' + '\n # \n The authors of this applet have not provided any information!'
+    }
     const { heading1, heading2, heading3, heading4, heading5, heading6, paragraph } = markdownStyle;
+    let alignment = 'center';
+
+    if (children.includes('::: hljs-right')) {
+      alignment = 'flex-end';
+    } else if (children.includes('::: hljs-left')) {
+      alignment = 'flex-start';
+    }
 
     return (
       <View
-        style={{ justifyContent: 'center', alignItems: 'center', marginHorizontal: 10}}
+        style={{ justifyContent: alignment, alignItems: alignment, marginHorizontal: 10 }}
       >
         <Markdown
-          style={{ heading1, heading2, heading3, heading4, heading5, heading6, paragraph }}
-          mergeStyle={ true }
+          style={{ heading1, heading2, heading3, heading4, heading5, heading6, paragraph, text: { flexDirection: 'row' } }}
+          mergeStyle={true}
           onLinkPress={(url) => {
             Linking.openURL(url).catch(error => console.warn('An error occurred: ', error));
           }}
@@ -137,3 +179,40 @@ MarkdownScreen.propTypes = {
 };
 
 export { MarkdownScreen };
+
+
+const checkNodeContent = (content) => {
+  content = content.replace(/(<([^>]+)>)/ig, '');
+
+  if (regex.test(content.trim())) content = content.trim().replace(/==/g, "")
+  if (content.indexOf("^") > -1 && content.indexOf('^^') === -1) return checkSuperscript(content)
+  if (content.indexOf("~") > -1 && content.indexOf('~~') === -1) return checkSubscript(content)
+
+  return content;
+}
+
+const checkSuperscript = (content) => {
+  if (content.indexOf("^") > -1 && content.indexOf('^^') === -1) {
+    return content.split("^").map((val, i) => {
+      if (i % 2 !== 0 && val.length > 0) {
+        return <Text style={{ fontSize: 13, lineHeight: 18 }}>{val}</Text>
+      } else {
+        return checkSubscript(val)
+      }
+    });
+  }
+  return content;
+}
+
+const checkSubscript = (content) => {
+  if (content.indexOf("~") > -1 && content.indexOf('~~') === -1) {
+    return content.split("~").map((val, i) => {
+      if (i % 2 !== 0 && val.length > 0) {
+        return <Text style={{ fontSize: 13, lineHeight: 18, textAlignVertical: "bottom" }}>{val}</Text>
+      } else {
+        return checkSuperscript(val)
+      }
+    })
+  }
+  return content;
+}