+++ app/scenes/Activity/index.js
@@ -126,7 +126,7 @@ class Activity extends React.Component {
     }
 
     if ((autoAdvance || fullScreen) && !optionalText) {
-      if (next === -1 && activity.compute && !isSummaryScreen) {
+      if (next === -1 && activity.compute && !activity.summaryDisabled && !isSummaryScreen) {
         this.setState({ isSummaryScreen: true });
         setSummaryScreen(true);
       } else {
@@ -281,6 +281,7 @@ class Activity extends React.Component {
                 if (
                   getNextPos(currentScreen, itemVisibility) === -1 &&
                   activity.compute &&
+                  !activity.summaryDisabled &&
                   !isSummaryScreen
                 ) {
                   this.setState({ isSummaryScreen: true });
@@ -319,7 +320,9 @@ class Activity extends React.Component {
             />
           </View>
         )}
-        {!fullScreen && <ActHeader title={activity.name.en} />}
+        {!fullScreen &&
+          <ActHeader title={activity.name.en} watermark={currentApplet.watermark} />
+        }
       </Container>
     );
   }