+++ app/scenes/ActivitySummary/index.js
@@ -1,30 +1,22 @@
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { getScoreFromResponse, evaluateScore, getMaxScore } from '../../services/scoring';
-import AsyncStorage from '@react-native-async-storage/async-storage';
 import { Parser } from 'expr-eval';
 import _ from 'lodash';
-
 import {
-  SafeAreaView,
   View,
   FlatList,
   StyleSheet,
   Dimensions,
-  StatusBar,
-  ImageBackground,
 } from 'react-native';
-import { Actions } from 'react-native-router-flux';
-import i18n from 'i18next';
+
 import { colors } from '../../themes/colors';
-import BaseText from '../../components/base_text/base_text';
-import { BodyText, Heading } from '../../components/core';
-import theme from '../../themes/base-theme';
-import FunButton from '../../components/core/FunButton';
-import { newAppletSelector } from '../../state/app/app.selectors';
+import { MarkdownScreen } from '../../components/core';
 import { parseAppletEvents } from '../../models/json-ld';
-import { setActivities } from '../../state/activities/activities.actions';
+import BaseText from '../../components/base_text/base_text';
+import { newAppletSelector } from '../../state/app/app.selectors';
+import { setActivities, setCumulativeActivities } from '../../state/activities/activities.actions';
+import { getScoreFromResponse, evaluateScore, getMaxScore } from '../../services/scoring';
 
 const { width } = Dimensions.get('window');
 const styles = StyleSheet.create({
@@ -54,25 +46,7 @@ const styles = StyleSheet.create({
   },
 });
 
-const DATA = [
-  {
-    category: i18n.t('activity_summary:category_suicide'),
-    message: i18n.t('activity_summary:category_suicide_message'),
-    score: '2.00',
-  },
-  {
-    category: i18n.t('activity_summary:category_emotional'),
-    message: i18n.t('activity_summary:category_emotional_message'),
-    score: '10.00',
-  },
-  {
-    category: i18n.t('activity_summary:sport'),
-    message: i18n.t('activity_summary:category_sport_message'),
-    score: '4.00',
-  },
-];
-
-const ActivitySummary = ({ responses, activity, applet, setActivities, activities }) => {
+const ActivitySummary = ({ responses, activity, applet, cumulativeActivities, setCumulativeActivities }) => {
   const [messages, setMessages] = useState([]);
 
   useEffect(() => {
@@ -108,31 +82,40 @@ const ActivitySummary = ({ responses, activity, applet, setActivities, activitie
     }, {});
 
     const reportMessages = [];
-    activity.messages.forEach(async (msg) => {
+    let cumActivities = [];
+    activity.messages.forEach(async (msg, i) => {
       const { jsExpression, message, outputType, nextActivity } = msg;
       const variableName = jsExpression.split(/[><]/g)[0];
       const category = variableName.trim().replace(/\s/g, '__');
       const expr = parser.parse(category + jsExpression.substr(variableName.length));
 
-      if (expr.evaluate(cumulativeScores)) {
-        const score = outputType == 'percentage' ? Math.round(cumulativeMaxScores[category] ? cumulativeScores[category] * 100 / cumulativeMaxScores[category] : 0) : cumulativeScores[category];
-        if (nextActivity)
-          AsyncStorage.setItem(`${activity.id}/nextActivity`, nextActivity)
+      const variableScores = {
+        [category]: outputType == 'percentage' ? Math.round(cumulativeMaxScores[category] ? cumulativeScores[category] * 100 / cumulativeMaxScores[category] : 0) : cumulativeScores[category]
+      }
+
+      if (expr.evaluate(variableScores)) {
+        if (nextActivity) cumActivities.push(nextActivity);
 
         reportMessages.push({
           category,
           message,
-          score: outputType == 'percentage' ? score + '%' : score
+          score: variableScores[category] + (outputType == 'percentage' ? '%' : ''),
         });
       }
     });
 
+    if (cumulativeActivities[`${activity.id}/nextActivity`]) {
+      cumActivities = _.difference(cumActivities, cumulativeActivities[`${activity.id}/nextActivity`]);
+      if (cumActivities.length > 0) {
+        cumActivities = [...cumulativeActivities[`${activity.id}/nextActivity`], ...cumActivities];
+        setCumulativeActivities({ [`${activity.id}/nextActivity`]: cumActivities });
+      }
+    } else {
+      setCumulativeActivities({ [`${activity.id}/nextActivity`]: cumActivities });
+    }
+
     setMessages(reportMessages);
   }, [responses]);
-
-  const onClose = () => {
-    Actions.push('activity_thanks');
-  };
 
   const renderItem = ({ item }) => {
     return (
@@ -143,7 +126,7 @@ const ActivitySummary = ({ responses, activity, applet, setActivities, activitie
         <BaseText style={{ fontSize: 24, color: colors.tertiary, paddingBottom: 20 }}>
           {item.score}
         </BaseText>
-        <BaseText style={{ fontSize: 15 }}>{item.message}</BaseText>
+        <MarkdownScreen>{item.message}</MarkdownScreen>
       </View>
     );
   };
@@ -171,10 +154,12 @@ ActivitySummary.propTypes = {
 const mapStateToProps = (state) => ({
   applet: newAppletSelector(state),
   activities: state.activities.activities,
+  cumulativeActivities: state.activities.cumulativeActivities,
 })
 
 const mapDispatchToProps = {
-  setActivities
+  setActivities,
+  setCumulativeActivities
 };
 
 export default connect(