+++ app/state/activities/activities.reducer.js
@@ -1,15 +1,20 @@
 
 export const initialState = {
-  activities: []
+  activities: [],
+  cumulativeActivities: {}
 };
 
 export default (state = initialState, action = {}) => {
-  console.log(action);
   switch (action.type) {
     case "SET_ACTIVITIES":
       return {
         ...state,
         activities: action.payload,
+      };
+    case "CUMULATIVE_ACTIVITIES":
+      return {
+        ...state,
+        cumulativeActivities: { ...state.cumulativeActivities, ...action.payload },
       };
 
     default: