+++ ios/MDCApp/AppDelegate.m
@@ -31,7 +31,7 @@
 
   if (@available(iOS 14, *)) {
     UIDatePicker *picker = [UIDatePicker appearance];
-    // picker.preferredDatePickerStyle = UIDatePickerStyleWheels;
+    picker.preferredDatePickerStyle = UIDatePickerStyleWheels;
   }
 
   rootView.backgroundColor = [[UIColor alloc] initWithRed:1.0f green:1.0f blue:1.0f alpha:1];